FIX: Installed CMake modules point to source dir
Some of the CMake modules go into CMAKE_BINARY_DIR with references to
CMAKE_BINARY_DIR so that they can be used mid-build. This makes building
and testing work, but once we install those modules to the install path,
they still point to CMAKE_BINARY_DIR, which we don't expect to continue
to exist in deployment, making the CMake modules break once deployed.

We can't fix this by just writing the CMake modules correctly in the
first place, to point to the install path, because the install path
won't be populated until after the build, and we need to use the modules
mid-build.

Instead, I opted to fix this by replacing refs to the binary dir with
the install path while installing the modules.